Initial upload of data files for LoRa paper
Data from two longer-term LoRa deployments.  The two farms are near Heidelberg in south-western Germany. The 1 km radio path for the strawberry farm setting is none line of sight passing agricultural areas, a few houses and some trees. The tomato farm is a complex radio environment with lots of greenhouses built of glass and steel. Distances are only between 100 and 200 m, but the receiver was placed in an office on the ground floor not facing to the direction of the field nodes. 3 of the field nodes were placed outdoors and 1 inside a greenhouse.
See trialsInfo.docx for details